pyproject.toml
[project]
name = "padelos-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "mcp>=1.25.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
]
